--- a/pyproject.toml
+++ b/pyproject.toml
@@ -37,9 +37,6 @@ classifiers = [
     "Typing :: Typed"
 ]
 requires-python = ">=3.8"
-dependencies = [
-    "aiofiles>=0.8.0",
-]
 
 [project.optional-dependencies]
 dev = [