--- a/pyproject.toml
+++ b/pyproject.toml
@@ -37,6 +37,7 @@ classifiers = [
     "Typing :: Typed"
 ]
 requires-python = ">=3.8"
+dependencies = []
 
 [project.optional-dependencies]
 dev = [