--- a/src/simpleenvs/filestream/__init__.py
+++ b/src/simpleenvs/filestream/__init__.py
@@ -1,0 +1,1 @@
+from .FileReader import GILOptimizedFileReader, read_env_file_optimized